requirements.txt
absl-py==2.1.0
altair==5.2.0
asgiref==3.7.2
astunparse==1.6.3
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.8.0rc2
distlib==0.3.8
Django==5.0.2
django-mapbox-location-field==2.0.0
dm-tree==0.1.8
filelock==3.13.1
flatbuffers==23.5.26
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.42
google-pasta==0.2.0
grpcio==1.62.0
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==3.0.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==4.2.0
protobuf==4.25.3
pyarrow==15.0.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyotp==2.9.0
pypng==0.20220715.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
qrcode==7.4.2
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.18.0
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.1.1
six==1.16.0
smmap==5.0.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
streamlit==1.31.1
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.0rc0
termcolor==2.4.0
threadpoolctl==3.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
validators==0.22.0
virtualenv==20.25.0
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
zipp==3.17.0
